# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Xoycoin Exchange",
    "description": "Métodos de la API de Xoycoin Exchange",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api-xoycoin.xellcoders.com",
      "description": "Xoycoin Exchange API"
    }
  ],
...
  "paths": {
...
    "/companies": {
      "put": {
        "tags": [
          "companies"
        ],
        "summary": "Update company",
        "description": "Actualiza la información de una empresa.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/companies_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Company object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                },
                "example": {
                  "id": "a860b92d-093d-4fe7-a7d5-e810cb7fcd3a",
                  "name\"": "Company name",
                  "status\"": "LOCKED"
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "companies"
        ],
        "summary": "Create company",
        "description": "Crea una empresa.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/companies_body_1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Company object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                },
                "example": {
                  "id": "a860b92d-093d-4fe7-a7d5-e810cb7fcd3a",
                  "name": "Company name",
                  "status": "LOCKED"
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/companies/{id}": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Get company by id",
        "description": "Obtiene la información de una empresa mediante el id proporcionado.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the company to return",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                },
                "example": {
                  "id": "a860b92d-093d-4fe7-a7d5-e810cb7fcd3a",
                  "name": "Company name",
                  "planType": "BASIC"
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "500": {
            "description": "ERROR: Company not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/deposits_companies/wallet/{walletId}": {
      "get": {
        "tags": [
          "deposits_companies"
        ],
        "summary": "Get company deposits by wallet id",
        "description": "Obtiene los depósitos de una empresa mediante el id de la wallet proporcionado.",
        "parameters": [
          {
            "name": "walletId",
            "in": "path",
            "description": "Id of the wallet to return deposits",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deposits array",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositCompany"
                },
                "example": [
                  {
                    "id": "b3de0177-020a-11ec-8efe-0242ac110003",
                    "walletId": "2c7d0d20-1cb5-4b21-b7e6-161480293c53",
                    "hash": "0x7ce560fa06ca34f0edf0b8922ee480a92cd62c46f85b9d1ca6743975624556df",
                    "sender": "mrd2zdpLJrdT6uVJKK6qkFz9FcPorbqbkc",
                    "amount": "1.0",
                    "fee": "0.001",
                    "confirmations": 1,
                    "status": "PENDING"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "500": {
            "description": "ERROR: Company not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/deposits_users/wallet/{walletId}": {
      "get": {
        "tags": [
          "deposits_users"
        ],
        "summary": "Get user deposits by wallet id",
        "description": "Obtiene los depósitos de un usuario mediante el id de la wallet proporcionado.",
        "parameters": [
          {
            "name": "walletId",
            "in": "path",
            "description": "Id of the wallet to return deposits",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array deposits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositUser"
                },
                "example": [
                  {
                    "id": "b3de0177-020a-11ec-8efe-0242ac110003",
                    "walletId": "2c7d0d20-1cb5-4b21-b7e6-161480293c53",
                    "hash": "0x7ce560fa06ca34f0edf0b8922ee480a92cd62c46f85b9d1ca6743975624556df",
                    "sender": "mrd2zdpLJrdT6uVJKK6qkFz9FcPorbqbkc",
                    "amount": "1.0",
                    "fee": "0.001",
                    "confirmations": 1,
                    "status": "PENDING"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "500": {
            "description": "ERROR: User not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/wallets_companies": {
      "post": {
        "tags": [
          "wallets_companies"
        ],
        "summary": "Create wallet company",
        "description": "Crea la wallet de una empresa.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/wallets_companies_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Wallet company object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletCompany"
                },
                "example": {
                  "id": "2c7d0d20-1cb5-4b21-b7e6-161480293c53",
                  "companyId": "458fba7d-ed81-4da8-a00c-21b5f520aee4",
                  "currency": "BTC",
                  "address": "mmqQY79NKhfJwoNwW4znsfATugPU1FaWU2",
                  "balance": "1.0"
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/wallets_companies/company/{companyId}": {
      "get": {
        "tags": [
          "wallets_companies"
        ],
        "summary": "Get wallets by company id",
        "description": "Obtiene las wallets de una empresa mediante el id proporcionado.",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "description": "Id of the company to return wallets",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Wallets array",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletCompany"
                },
                "example": [
                  {
                    "id": "2c7d0d20-1cb5-4b21-b7e6-161480293c53",
                    "companyId": "458fba7d-ed81-4da8-a00c-21b5f520aee4",
                    "currency": "BTC",
                    "address": "mmqQY79NKhfJwoNwW4znsfATugPU1FaWU2",
                    "balance": "1.0"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "500": {
            "description": "ERROR: Company not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/withdrawals_companies/wallet/{walletId}": {
      "get": {
        "tags": [
          "withdrawals_companies"
        ],
        "summary": "Get withdrawals by wallet id",
        "description": "Obtiene los retiros/ordenes de una empresa mediante el id de la wallet proporcionado.",
        "parameters": [
          {
            "name": "walletId",
            "in": "path",
            "description": "Id of the wallet to return withdrawals",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Withdrawals array",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalCompany"
                },
                "example": [
                  {
                    "id": "b3de0177-020a-11ec-8efe-0242ac110003",
                    "walletId": "2c7d0d20-1cb5-4b21-b7e6-161480293c53",
                    "hash": "0x7ce560fa06ca34f0edf0b8922ee480a92cd62c46f85b9d1ca6743975624556df",
                    "receiver": "mrd2zdpLJrdT6uVJKK6qkFz9FcPorbqbkc",
                    "amount": "1.0",
                    "fee": "0.001",
                    "confirmations": 1,
                    "status": "PENDING"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "500": {
            "description": "ERROR: Cannot get withdrawals"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/withdrawals_users/wallet/{walletId}": {
      "get": {
        "tags": [
          "withdrawals_users"
        ],
        "summary": "Get withdrawals by wallet id",
        "description": "Obtiene los retiros/ordenes de un usuario mediante el id de la wallet proporcionado.",
        "parameters": [
          {
            "name": "walletId",
            "in": "path",
            "description": "Id of the wallet to return withdrawals",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Withdrawals array",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalUser"
                },
                "example": [
                  {
                    "id": "b3de0177-020a-11ec-8efe-0242ac110003",
                    "walletId": "2c7d0d20-1cb5-4b21-b7e6-161480293c53",
                    "hash": "0x7ce560fa06ca34f0edf0b8922ee480a92cd62c46f85b9d1ca6743975624556df",
                    "receiver": "mrd2zdpLJrdT6uVJKK6qkFz9FcPorbqbkc",
                    "amount": "1.0",
                    "fee": "0.001",
                    "confirmations": 1,
                    "status": "PENDING"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "500": {
            "description": "ERROR: Cannot get withdrawals"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Company": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "planType": {
            "type": "string"
          }
        }
      },
...
      "WalletCompany": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "balance": {
            "type": "string"
          }
        }
      },
...
      "DepositCompany": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "walletId": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "sender": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "fee": {
            "type": "string"
          },
          "confirmations": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "WithdrawalCompany": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "walletId": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "sender": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "fee": {
            "type": "string"
          },
          "confirmations": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "DepositUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "walletId": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "sender": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "fee": {
            "type": "string"
          },
          "confirmations": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "WithdrawalUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "walletId": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "sender": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "fee": {
            "type": "string"
          },
          "confirmations": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "companies_body": {
        "required": [
          "id",
          "name",
          "planType"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "planType": {
            "type": "string"
          }
        }
      },
      "companies_body_1": {
        "required": [
          "name",
          "planType"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "planType": {
            "type": "string"
          }
        }
      },
...
      "wallets_companies_body": {
        "required": [
          "companyId",
          "currency"
        ],
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}